Moved data types to `rt_model` and `cfg_model`.
