+++ src/lib.rs
@@ -2,17 +2,9 @@
 
 //! Automation that starts where it stops.
 
-pub use crate::{
-    destination::Destination, stations::Stations, train::Train, visit_fn::VisitFn,
-    visit_status::VisitStatus, workload::Workload,
-};
+pub use crate::train::Train;
 
 pub mod cfg_model;
 pub mod rt_model;
 
-mod destination;
-mod stations;
 mod train;
-mod visit_fn;
-mod visit_status;
-mod workload;
