+++ src/rt_model.rs
@@ -2,6 +2,11 @@
 //!
 //! Types in this module are analogous to build artifacts.
 
-pub use self::station::Station;
+pub use self::{
+    destination::Destination, station::Station, stations::Stations, visit_status::VisitStatus,
+};
 
+mod destination;
 mod station;
+mod stations;
+mod visit_status;
